requirements.txt
adjusttext==0.8
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antropy==0.1.6
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.2
arch==6.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.11.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.7.0
bottleneck==1.3.7
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
#chatstream==0.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.14
dill==0.3.8
distro==1.9.0
executing==0.8.3
faicons==0.2.2
fastapi==0.110.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.25.0
#freetype==2.12.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
gitpython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h2o==3.46.0.1
holidays==0.33
htmltools==0.5.1
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.3.1
ipykernel==6.28.0
ipython==8.18.0
ipywidgets==8.1.2
jedi==0.18.1
jinja2==3.1.3
joblib==1.2.0
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_events==0.8.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-widgets==3.0.10
jupyterlab_pygments==0.1.2
jupyterlab_server==2.25.1
kiwisolver==1.4.4
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.33
#langchain-coreadjusttext==0.8
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antropy==0.1.6
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.2
arch==6.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.11.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.7.0
bottleneck==1.3.7
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
#chatstream==0.1.5
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.14
dill==0.3.8
distro==1.9.0
executing==0.8.3
faicons==0.2.2
fastapi==0.110.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.25.0
#freetype==2.13.3
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
gitpython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h2o==3.46.0.1
holidays==0.33
htmltools==0.5.1
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.3.1
ipykernel==6.28.0
ipython==8.18.0
ipywidgets==8.1.2
jedi==0.18.1
jinja2==3.1.3
joblib==1.2.0
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_events==0.8.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-widgets==3.0.10
jupyterlab_pygments==0.1.2
jupyterlab_server==2.25.1
kiwisolver==1.4.4
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-experimental==0.0.57
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.27
#lerc==3.0
linkify-it-py==2.0.3
llvmlite==0.42.0
#lz4-c==1.9.4
markdown-it-py==3.0.0
markupsafe==2.1.3
marshmallow==3.21.1
matplotlib==3.8.0
#matplotlib-base==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==2.0.4
mizani==0.9.3
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
#ncurses==6.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook-shim==0.2.3
numba==0.59.0
numexpr==2.8.7
numpy==1.26.4
#numpy-base==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.14.1
#openjpeg==2.4.0
#openssl==3.0.13
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandas-flavor==0.6.0
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.2
patsy==0.5.6
pdfkit==1.0.0
pexpect==4.8.0
pillow==10.2.0
pip==23.3.1
platformdirs==3.10.0
plotly==5.20.0
plotnine==0.12.4
polars==0.20.16
posthog==3.5.0
pox==0.3.4
ppft==1.7.6.8
prometheus_client==0.14.1
#prompt-toolkit==3.0.36
#prompt_toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pulsar-client==3.4.0
pure_eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
pydeck==0.8.1b0
pygments==2.15.1
pyparsing==3.0.9
pypdf==4.2.0
pypika==0.48.9
pyproject-hooks==1.0.0
pysocks==1.7.1
#python==3.10.13
python-dateutil==2.8.2
python-dotenv==1.0.1
#python-fastjsonschema==2.16.2
python-json-logger==2.0.7
python-multipart==0.0.9
#python-tzdata==2023.3
pytimetk==0.4.0
pytube==15.0.0
pytz==2023.3.post1
pyyaml==6.0.1
pyzmq==25.1.2
quarto==0.1.0
questionary==2.0.1
#readline==8.2
referencing==0.30.2
regex==2023.12.25
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
ridgeplot==0.1.24
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.13.2
send2trash==1.8.2
sentence-transformers==2.5.1
setuptools==68.2.2
shiny==0.8.1
shinywidgets==0.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.28
#sqlite==3.41.2
#sqlite-vss==0.1.2
stack_data==0.2.0
starlette==0.36.3
statsmodels==0.14.1
#stochastic==0.7.0
streamlit==1.32.2
supersmoother==0.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==2.2.0
tiktoken==0.6.0
tinycss2==1.2.1
#tk==8.6.12
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.7.1
transformers==4.38.2
tsfeatures==0.4.5
typer==0.9.0
typing-extensions==4.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
#tzdata==2024a
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.28.0
#uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
websockets==12.0
#wheel==0.41.2
widgetsnbextension==4.0.10
wrapt==1.16.0
xarray==2023.10.1
#xz==5.4.6
#yaml==0.2.5
#yarl==1.9.4
youtube-transcript-api==0.6.2
#zeromq==4.3.5
zipp==3.18.1
#zlib==1.2.13
#zstd==1.5.5
